scripts/requirements.txt
# Core dependencies
pyyaml>=6.0

# AI Providers (install only what you need)
# For Google Vertex AI
google-cloud-aiplatform>=1.38.0

# For OpenAI
openai>=1.3.0

# For Anthropic Claude
anthropic>=0.7.0

# For local Ollama (just uses requests)
requests>=2.31.0

# Vector store (optional - only needed for embedding features)
qdrant-client>=1.7.0

# Additional utilities
python-dateutil>=2.8.0
